Finish application logic

+++ CsvParserToDb.Domain.Shared/Entities/TripEntity.cs
@@ -6,7 +6,7 @@ public class TripEntity
     public DateTime TpepPickupDatetime { get; set; }
     public DateTime TpepDropoffDatetime { get; set; }
     public int PassengerCount { get; set; }
-    public float TripDistance { get; set; }
+    public double TripDistance { get; set; }
     public string StoreAndFwdFlag { get; set; }
     public int PULocationID { get; set; }
     public int DOLocationID { get; set; }
